Rename core-branding into ui-core-branding
Following the name convensions
