+++ packages/dotcom-ui-core-branding/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@financial-times/dotcom-core-branding",
+  "name": "@financial-times/dotcom-ui-core-branding",
   "version": "0.0.0",
   "description": "",
   "main": "component.js",
@@ -28,7 +28,7 @@
   "repository": {
     "type": "git",
     "repository": "https://github.com/Financial-Times/dotcom-page-kit.git",
-    "directory": "packages/dotcom-core-branding"
+    "directory": "packages/dotcom-ui-core-branding"
   },
-  "homepage": "https://github.com/Financial-Times/dotcom-page-kit/tree/master/packages/dotcom-core-branding"
+  "homepage": "https://github.com/Financial-Times/dotcom-page-kit/tree/master/packages/dotcom-ui-core-branding"
 }
