+++ packages/dotcom-ui-core-branding/src/__test__/components/Favicon.test.tsx
