+++ packages/dotcom-ui-core-branding/src/__test__/components/SocialMedia.test.tsx
