+++ packages/dotcom-ui-core-branding/src/components/Favicon.tsx
