+++ packages/dotcom-ui-core-branding/src/defaultColors.ts
