+++ packages/dotcom-ui-core-branding/src/lib/imageServiceIconURL.ts
