+++ packages/dotcom-ui-layout/package.json
@@ -25,7 +25,7 @@
     "@financial-times/dotcom-ui-header": "file:../dotcom-ui-header",
     "@financial-times/dotcom-ui-footer": "file:../dotcom-ui-footer",
     "@financial-times/dotcom-types-navigation": "file:../dotcom-types-navigation",
-    "@financial-times/dotcom-core-branding": "file:../dotcom-core-branding",
+    "@financial-times/dotcom-ui-core-branding": "file:../dotcom-ui-core-branding",
     "focus-visible": "^4.1.5"
   },
   "peerDependencies": {
