+++ packages/dotcom-ui-shell/package.json
@@ -24,7 +24,7 @@
     "@financial-times/dotcom-ui-app-context": "file:../dotcom-ui-app-context",
     "@financial-times/dotcom-ui-flags": "file:../dotcom-ui-flags",
     "@financial-times/dotcom-ui-polyfill-service": "file:../dotcom-ui-polyfill-service",
-    "@financial-times/dotcom-core-branding": "file:../dotcom-core-branding"
+    "@financial-times/dotcom-ui-core-branding": "file:../dotcom-ui-core-branding"
   },
   "peerDependencies": {
     "react": "^16.8.6"
